% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-level-mapping.R
\name{gisa_auto_map_levels}
\alias{gisa_auto_map_levels}
\title{Map auto categorical variable levels}
\usage{
gisa_auto_map_levels(data)
}
\arguments{
\item{data}{Data frame with tidy names.}
}
\description{
Map auto categorical variable levels
}
