% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{gisa_extract_triangle}
\alias{gisa_extract_triangle}
\title{Extract triangle}
\usage{
gisa_extract_triangle(
  data,
  type = c("paid", "incurred", "count"),
  minor_coverage_type,
  evaluation_date = ymd("2018-12-31")
)
}
\arguments{
\item{data}{A loss development exhibit table.}

\item{type}{One of "paid", "incurred", or "count".}

\item{minor_coverage_type}{Minor coverage type.}

\item{evaluation_date}{Evaluation date.}
}
\description{
Extract triangle
}
