% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-level-mapping.R
\name{gisa_liab_map_levels}
\alias{gisa_liab_map_levels}
\title{Map liability categorical variable levels}
\usage{
gisa_liab_map_levels(data)
}
\arguments{
\item{data}{Data frame with tidy names.}
}
\description{
Map liability categorical variable levels
}
\keyword{internal}
