% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{gisa_origin_dev}
\alias{gisa_origin_dev}
\title{Compute origin and development periods}
\usage{
gisa_origin_dev(data, accident_col, entry_col)
}
\arguments{
\item{data}{Data frame with loss development data.}

\item{accident_col}{Column corresponding to accident half year, as an unquoted string.}

\item{entry_col}{Column corresponding to entry half year, as an unquoted string.}
}
\description{
Compute origin and development periods
}
