% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{gisa_process}
\alias{gisa_process}
\alias{gisa_process_clsp}
\alias{gisa_process_auto_dev}
\alias{gisa_process_auto_cat}
\title{Process GISA data}
\usage{
gisa_process_clsp(path)

gisa_process_auto_dev(path)

gisa_process_auto_cat(path)
}
\arguments{
\item{path}{Path to directory of CSV files.}
}
\value{
A list of tibbles whose names denote GISA exhibits.
}
\description{
Process GISA data
}
