% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-mapping.R
\name{gisa_select_cols}
\alias{gisa_select_cols}
\title{Select columns based on format number}
\usage{
gisa_select_cols(data, format_number)
}
\arguments{
\item{data}{Data frame}

\item{format_number}{GISA format number}
}
\description{
Select columns based on format number
}
