% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip.R
\name{gisa_unzip}
\alias{gisa_unzip}
\title{Extract zip files}
\usage{
gisa_unzip(zips_dir, extract_dir = tempdir())
}
\arguments{
\item{zips_dir}{The directory of the zip files.}

\item{extract_dir}{The directory to extract to.}
}
\description{
Extract files, up to two layers of .zip files, from a directory into a
specified directory.
}
