% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{gisa_extract_triangle}
\alias{gisa_extract_triangle}
\title{Extract triangle}
\usage{
gisa_extract_triangle(data, type = c("paid", "incurred"))
}
\arguments{
\item{data}{A loss development exhibit table.}

\item{type}{Either "paid" or "incurred".}
}
\description{
Extract triangle
}
